Added example and fixes for persistant logging mosly code from https://github.com/serilog/serilog-sinks-seq
