+++ sample/Serilog.Sinks.Elasticsearch.Sample/appsettings.json
@@ -1,0 +1,5 @@
+{
+    "ConnectionStrings": {
+        "elasticsearch": "http://elastic:changeme@localhost:9200"
+    }
+}
