+++ src/Serilog.Sinks.Elasticsearch/Sinks/ElasticSearch/Durable/FileSet.cs
@@ -24,6 +24,9 @@ using Serilog.Debugging;
 
 namespace Serilog.Sinks.Elasticsearch.Durable
 {
+    /// <summary>
+    /// https://github.com/serilog/serilog-sinks-seq/blob/v4.0.0/src/Serilog.Sinks.Seq/Sinks/Seq/Durable/FileSet.cs
+    /// </summary>
     class FileSet
     {
         readonly string _bookmarkFilename;
